deploy/requirements.txt
numpy==1.19.3
opencv-python==4.5.1.48
Pillow==8.1.2
streamlit==0.81.1
torch==1.8.1
